// src/account/types.ts
export interface DeployAccountPayload {
  /** Account class hash */
  classHash: string;
  /** Constructor calldata */
  constructorCalldata?: Felt252Input[];
  /** Contract address salt */
  addressSalt?: string;
}